+++ client/src/components/gather/Gather.js
@@ -14,6 +14,20 @@ function Gather() {
       >
         군적금 신청
       </button>
+      <button
+        onClick={() => {
+          history("/gather/add-goal");
+        }}
+      >
+        목표설정
+      </button>
+      <button
+        onClick={() => {
+          history("/gather/add-safebox");
+        }}
+      >
+        세이프박스
+      </button>
     </div>
   );
 }
