+++ client/src/components/gather/addGoal/StoreSvg.js
